Combining server.js and database.js and creating an edit lesson api call
