+++ server.js
@@ -46,7 +46,6 @@ app.post('/api/register', (req, res) => {
     let username = body.username;
     let password = body.password;
 
-
     //Hash Password
     bcrypt.genSalt(10, (err, salt) => {
         bcrypt.hash(password, salt, (err, hash) => {
@@ -123,3 +122,97 @@ app.post('/api/grade', (req, res) => {
 });
 
 app.listen(port, () => console.log(`Listening on port ${port}`));
+
+var sqlite = require('sqlite3').verbose();
+const DBSOURCE = "./client/src/database.db";
+
+module.exports = (app, db) =>{
+
+
+/************** Lesson Requests ****************/
+    //GET all Lessons
+    app.get('/api/Lesson/all', (req,res) => {
+        let sql = 'SELECT * FROM Lesson  ';
+        let params = [];
+        db.all(sql, params, (err, rows) =>{
+            if (err){
+                res.status(400).json({
+                    "error" : err.message,
+                    "message" : "Failure"});
+                return;
+            }
+            res.json({
+                message: "Success",
+                data: rows});
+        });
+    });
+
+    //GET single Lesson
+    app.get('/Lesson/:id', (req, res) => {
+        let sql = 'SELECT * FROM Lesson WHERE lesson_id = ?';
+        let lessonNum = [req.params.id];
+
+        //get query to database for lesson with :id
+        db.get(sql, lessonNum, (err, row) => {
+            if (err){
+                res.status(400).json({
+                    "error" : err.message,
+                    "message" : "Failure"});
+                return;
+            }
+            res.json({
+                message: "Success",
+                data: row});
+        });
+    });
+
+    app.post('/NewLesson/', (req, res) => {
+        // let sql = 'SELECT * FROM Lesson WHERE lesson_id = ?';
+        let sql = 'INSERT INTO Lesson VALUES ?';
+
+        let lessonID = 100;
+        let lessonName = [req.params.name];
+        let lessonQuestion = [req.params.question];
+        let lessonHint = [req.params.hint];
+
+        let values = [lessonID, lessonName, lessonQuestion, lessonHint]
+
+        
+
+        //get query to database for lesson with :id
+        db.get(sql, values, (err, row) => {
+            if (err){
+                res.status(400).json({
+                    "error" : err.message,
+                    "message" : "Failure"});
+                return;
+            }
+            res.json({
+                message: "Success",
+                data: row
+            });
+        });
+    });
+
+/****************** User Requests *****************/
+
+    app.get('/User/:userid', (req,res) => {
+        let sql = 'SELECT * FROM User WHERE user_id = ?';
+
+        db.get(sql, req.params.userid, (err,row) => {
+            if(err){
+                res.status(400).json({
+                    "error": err.message,
+                    "message" : "Failure"})
+                    return;
+            }
+            res.json({
+                message: "Success",
+                data: row});
+        });
+    });
+    //Only time we need to close database is on SIGINT
+//    process.on('SIGINT', () =>{
+//        db.close();
+//    });
+};
